# Optimizing Stacked Borrows (part 2): Shrink Item