--- a/tests/fail/stacked_borrows/deallocate_against_barrier1.rs
+++ b/tests/fail/stacked_borrows/deallocate_against_barrier1.rs
@@ -1,4 +1,4 @@
-//@error-pattern: deallocating while item is protected
+//@error-pattern: deallocating while item
 
 fn inner(x: &mut i32, f: fn(&mut i32)) {
     // `f` may mutate, but it may not deallocate!