… src/stacked_borrows/item.rs
… tests/fail/stacked_borrows/aliasing_mut2.stderr
… tests/fail/stacked_borrows/aliasing_mut4.stderr
… tests/fail/stacked_borrows/deallocate_against_barrier1.stderr
… tests/fail/stacked_borrows/deallocate_against_barrier2.rs
… tests/fail/stacked_borrows/deallocate_against_barrier2.stderr
… tests/fail/stacked_borrows/illegal_write6.stderr
… tests/fail/stacked_borrows/invalidate_against_barrier1.stderr
… tests/fail/stacked_borrows/invalidate_against_barrier2.stderr
… tests/fail/stacked_borrows/newtype_retagging.rs
… tests/fail/stacked_borrows/newtype_retagging.stderr