@RalfJung 2022-07-03
> This moves protectors out of Item, storing them both in a global HashSet which contains all currently-protected tags as well as a Vec<SbTag> on each Frame so that when we return from a function we know which tags to remove from the protected set.

How is that a correct semantics? Note that with a type like `&(i32, Cell<i32>)`, since https://github.com/rust-lang/miri/pull/2248 landed this tag is protected for the `i32` part of the memory but not for the `Cell` part.