@saethlin 2022-07-03
Yes this absolutely implements the wrong semantics right now, and the two tests for internal mutability fail. I'm going to touch up the rest of this first and hope you decide on the behavior of `UnsafeCell` in the meantime.