@RalfJung 2022-07-03
Well I am personally pretty much settled that I want them to completely infect the reference.

But this seems like a big change and I didn't see a lot of other people (in particular the lang team) enthusiastically agree with that...