@RalfJung 2022-07-03
Meanwhile, what about the following scheme: we sacrifice another bit in the item to indicate whether or not this item is protected. That should be enough, because all items for a given tag `t` are *either* protected by the same CallId, or not protected; i.e., we never have two items with the same tag protected by two different CallIds.