@saethlin 2022-07-03
> in particular the lang team

:shrug: All that I can find related to this decision is [the rfcbot comment](https://github.com/rust-lang/rust/pull/98017#issuecomment-1163625324) which got un-nominated with no evidence of discussion. I haven't seen any objections from any lang team members, and it's only been 5 days since it was un-nominated. Maybe I'm just very patient.

I was considering sacrificing another bit anyway to bypass the check in `item_popped`. But I'm not sure that resolves the ambiguity, don't we still have the problem that for some Tag, it will be protected in the stacks that are not within an UnsafeCell, and it will not be protected in the stacks that are inside an UnsafeCell?