@saethlin 2022-07-03
Oh :facepalm: wait I get it. Yes this is a great idea.