@RalfJung 2022-07-03
>  All that I can find related to this decision is https://github.com/rust-lang/rust/pull/98017#issuecomment-1163625324 which got un-nominated with no evidence of discussion. I haven't seen any objections from any lang team members, and it's only been 5 days since it was un-nominated. Maybe I'm just very patient.

That's a different discussion. If/when that lands (which I expect it will), that brings the doc in sync with what Miri already does since https://github.com/rust-lang/miri/pull/2248.

What you would have needed is https://github.com/rust-lang/unsafe-code-guidelines/issues/236.
But anyway we seem to agree that compression is possible even with the current semantics. :)