@saethlin 2022-07-03
Now I just have UI test failures because printing a tag can no longer print the CallId that it was protected by. Is that okay?