@saethlin 2022-07-03
Also, should we even have a second `Item` struct, or just use the packed one everywhere? I don't think there's any performance argument to be made, but it would make the codebase simpler overall.