@RalfJung 2022-07-04
> Also, should we even have a second `Item` struct, or just use the packed one everywhere? I don't think there's any performance argument to be made, but it would make the codebase simpler overall.

Only having the packed thing would make the codebase simpler? Sure, then go for it. (I guess for `struct` that is usually the case; when packing an `enum` one really wants the unpacked type to do `match`.)