@RalfJung 2022-07-12
Yes it does. :)
@rustbot ready